fix(java): Don't generate pagination with nonempty path

+++ seed/java-sdk/object/src/main/java/com/seed/object/types/Type.java
@@ -65,11 +65,11 @@ public final class Type {
 
     private final Name nineteen;
 
-    private final int twenty;
+    private final long twenty;
 
     private final long twentyone;
 
-    private final double twentytwo;
+    private final float twentytwo;
 
     private final BigInteger twentythree;
 
@@ -94,9 +94,9 @@ public final class Type {
             List<Map<String, Integer>> sixteen,
             List<Optional<UUID>> seventeen,
             Name nineteen,
-            int twenty,
+            long twenty,
             long twentyone,
-            double twentytwo,
+            float twentytwo,
             BigInteger twentythree,
             Map<String, Object> additionalProperties) {
         this.one = one;
@@ -220,7 +220,7 @@ public final class Type {
     }
 
     @JsonProperty("twenty")
-    public int getTwenty() {
+    public long getTwenty() {
         return twenty;
     }
 
@@ -230,7 +230,7 @@ public final class Type {
     }
 
     @JsonProperty("twentytwo")
-    public double getTwentytwo() {
+    public float getTwentytwo() {
         return twentytwo;
     }
 
@@ -358,7 +358,7 @@ public final class Type {
     }
 
     public interface TwentyStage {
-        TwentyoneStage twenty(int twenty);
+        TwentyoneStage twenty(long twenty);
     }
 
     public interface TwentyoneStage {
@@ -366,7 +366,7 @@ public final class Type {
     }
 
     public interface TwentytwoStage {
-        TwentythreeStage twentytwo(double twentytwo);
+        TwentythreeStage twentytwo(float twentytwo);
     }
 
     public interface TwentythreeStage {
@@ -457,11 +457,11 @@ public final class Type {
 
         private Name nineteen;
 
-        private int twenty;
+        private long twenty;
 
         private long twentyone;
 
-        private double twentytwo;
+        private float twentytwo;
 
         private BigInteger twentythree;
 
@@ -590,7 +590,7 @@ public final class Type {
 
         @java.lang.Override
         @JsonSetter("twenty")
-        public TwentyoneStage twenty(int twenty) {
+        public TwentyoneStage twenty(long twenty) {
             this.twenty = twenty;
             return this;
         }
@@ -604,7 +604,7 @@ public final class Type {
 
         @java.lang.Override
         @JsonSetter("twentytwo")
-        public TwentythreeStage twentytwo(double twentytwo) {
+        public TwentythreeStage twentytwo(float twentytwo) {
             this.twentytwo = twentytwo;
             return this;
         }
